+++ src/Jobs/EdgePurgeUrls.php
@@ -26,6 +26,18 @@ class EdgePurgeUrls implements ShouldQueue, ShouldBeUnique
 
     public function handle(): void
     {
+        \Log::info('Purging URLs: ' . implode(', ', $this->urls));
+
+        if (!TwillEdgePurge::canDispatchInvalidations()) {
+            \Log::info('Deferred: ' . TwillEdgePurge::canDispatchInvalidations());
+
+            $this->release(60); // seconds
+
+            return;
+        }
+
+        \Log::info('Execute Purge');
+
         TwillEdgePurge::purgeUrls($this->urls);
     }
 }
